+++ apps/website/src/docs/components/sidebar/sidebar.tsx
@@ -37,8 +37,7 @@ import {
   NavItemProps,
 } from '@saas-ui/react'
 
-import { ChevronRightIcon } from '@chakra-ui/icons'
-import { FiChevronDown } from 'react-icons/fi'
+import { FiChevronDown, FiChevronRight } from 'react-icons/fi'
 import { LuCheck } from 'react-icons/lu'
 import SaasUIGlyph from '@/components/saas-ui-glyph'
 import { NextjsIcon } from '@/components/logos/nextjs'
@@ -82,7 +81,8 @@ function SidebarHeader({ isOpen, isActive, children, ...props }: any) {
             },
           }}
         >
-          <ChevronRightIcon
+          <Icon
+            as={FiChevronRight}
             transform={isOpen && 'rotate(90deg)'}
             transition="transform .2s ease-in"
           />