+++ packages/saas-ui-charts/tsup.config.ts
@@ -7,13 +7,7 @@ export default defineConfig({
   },
   clean: true,
   sourcemap: true,
-  external: [
-    'react',
-    'recharts',
-    '@chakra-ui/react',
-    '@chakra-ui/system',
-    '@chakra-ui/utils',
-  ],
+  external: ['react', 'recharts', '@chakra-ui/react', '@chakra-ui/utils'],
   format: ['esm', 'cjs'],
   banner: {
     js: "'use client'",