+++ packages/saas-ui-auth/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @saas-ui/auth
 
+## 3.4.0
+
+### Minor Changes
+
+- 2ac496a: Upgraded to Chakra UI v2.9.3
+
+### Patch Changes
+
+- Updated dependencies [2ac496a]
+  - @saas-ui/forms@2.7.0
+  - @saas-ui/hooks@2.1.0
+  - @saas-ui/core@2.6.0
+
 ## 3.3.1
 
 ### Patch Changes