+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # saas-ui-storybook
 
+## 2.1.0
+
+### Minor Changes
+
+- 2ac496a: Upgraded to Chakra UI v2.9.3
+
+### Patch Changes
+
+- Updated dependencies [2ac496a]
+  - @saas-ui/storybook-addon@4.0.0
+  - @saas-ui/react@2.9.0
+
 ## 2.0.11
 
 ### Patch Changes